requirements.txt
click==6.7
Django==1.11.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
paho-mqtt==1.3.1
pytz==2017.3
Werkzeug==0.12.2
